update chapter 2
