+++ crates/dessins/src/params.rs
@@ -1,4 +1,4 @@
-use crate::{chapter_1, reflect::control_reflect, shapes::Shapes};
+use crate::{reflect::ControllableParams, shapes::Shapes};
 use nannou::prelude::*;
 
 pub struct DesignController {
@@ -8,11 +8,12 @@ pub struct DesignController {
 
 #[derive(PartialEq, Reflect)]
 pub enum DesignParams {
-    Polygon(chapter_1::Polygon),
-    Star(chapter_1::Star),
-    Composition1(chapter_1::Composition1),
-    Composition2(chapter_1::Composition2),
-    Jolygon(chapter_1::Jolygon),
+    Polygon(crate::chapter_1::Polygon),
+    Star(crate::chapter_1::Star),
+    Composition1(crate::chapter_1::Composition1),
+    Composition2(crate::chapter_1::Composition2),
+    Jolygon(crate::chapter_1::Jolygon),
+    Shape(crate::chapter_2::Params),
 }
 
 #[derive(PartialEq, Eq, Clone, Copy)]
@@ -22,6 +23,7 @@ pub enum DesignVariant {
     Composition1,
     Composition2,
     Jolygon,
+    Shape,
 }
 
 impl DesignController {
@@ -53,6 +55,9 @@ impl DesignController {
                 changed |= ui
                     .selectable_value(&mut self.selected, DesignVariant::Jolygon, "jolygon")
                     .changed();
+                changed |= ui
+                    .selectable_value(&mut self.selected, DesignVariant::Shape, "shape")
+                    .changed();
             });
         });
 
@@ -67,11 +72,16 @@ impl DesignController {
 impl DesignVariant {
     pub fn get_params(&self) -> DesignParams {
         match self {
-            Self::Polygon => DesignParams::Polygon(chapter_1::Polygon::default()),
-            Self::Star => DesignParams::Star(chapter_1::Star::default()),
-            Self::Composition1 => DesignParams::Composition1(chapter_1::Composition1::default()),
-            Self::Composition2 => DesignParams::Composition2(chapter_1::Composition2::default()),
-            Self::Jolygon => DesignParams::Jolygon(chapter_1::Jolygon::default()),
+            Self::Polygon => DesignParams::Polygon(crate::chapter_1::Polygon::default()),
+            Self::Star => DesignParams::Star(crate::chapter_1::Star::default()),
+            Self::Composition1 => {
+                DesignParams::Composition1(crate::chapter_1::Composition1::default())
+            }
+            Self::Composition2 => {
+                DesignParams::Composition2(crate::chapter_1::Composition2::default())
+            }
+            Self::Jolygon => DesignParams::Jolygon(crate::chapter_1::Jolygon::default()),
+            Self::Shape => DesignParams::Shape(crate::chapter_2::Params::default()),
         }
     }
 }
@@ -84,16 +94,18 @@ impl DesignParams {
             DesignParams::Composition1(params) => params.calculate_shapes(),
             DesignParams::Composition2(params) => params.calculate_shapes(),
             DesignParams::Jolygon(params) => params.calculate_shapes(),
+            DesignParams::Shape(params) => params.calculate_shapes(),
         }
     }
 
     pub fn control(&mut self, ctx: &mut egui::Context) -> (bool, Option<Color>) {
         match self {
-            DesignParams::Polygon(params) => control_reflect(params, ctx),
-            DesignParams::Star(params) => control_reflect(params, ctx),
-            DesignParams::Composition1(params) => control_reflect(params, ctx),
-            DesignParams::Composition2(params) => control_reflect(params, ctx),
-            DesignParams::Jolygon(params) => control_reflect(params, ctx),
+            DesignParams::Polygon(params) => params.control_params(ctx),
+            DesignParams::Star(params) => params.control_params(ctx),
+            DesignParams::Composition1(params) => params.control_params(ctx),
+            DesignParams::Composition2(params) => params.control_params(ctx),
+            DesignParams::Jolygon(params) => params.control_params(ctx),
+            DesignParams::Shape(params) => params.control_params(ctx),
         }
     }
 }
